+++ game.js
@@ -69,6 +69,7 @@ var Game = (function() {
         self.minerUI.update(delta);
         self.colonyUI.update(delta);
         self.laboratoryUI.update(delta);
+        self.goldenEventUI.update(delta);
 
         self.updateAutoSave(delta);
 
@@ -146,6 +147,7 @@ var Game = (function() {
         this.settings.save(data);
         this.miners.save(data);
         this.laboratory.save(data);
+        this.goldenEvents.save(data);
         this.interstellar.save(data);
         this.stargaze.save(data);
         this.updates.save(data);
@@ -169,6 +171,7 @@ var Game = (function() {
             this.buildings.load(data);
             this.miners.load(data);
             this.laboratory.load(data);
+            this.goldenEvents.load(data);
             this.stargaze.load(data);
             this.tech.load(data);
             this.interstellar.load(data); 
@@ -251,6 +254,7 @@ var Game = (function() {
         self.stargaze.initialise();
         self.miners.initialise();
         self.laboratory.initialise();
+        self.goldenEvents.initialise();
 
         // Now load
         self.load();
@@ -259,6 +263,7 @@ var Game = (function() {
         self.minerUI.initialise();
         self.colonyUI.initialise();
         self.laboratoryUI.initialise();
+        self.goldenEventUI.initialise();
 
         for(var i = 0; i < self.uiComponents.length; i++) {
             self.uiComponents[i].initialise();